requirements.txt
mcp>=0.1.0
aiomysql==0.2.0
python-dotenv>=0.19.0 
sqlparse>=0.4.2